+++ vector/src/main/java/im/vector/app/core/glide/VectorGlideModelLoader.kt
@@ -16,6 +16,7 @@
 
 package im.vector.app.core.glide
 
+import android.content.Context
 import com.bumptech.glide.Priority
 import com.bumptech.glide.load.DataSource
 import com.bumptech.glide.load.Options
@@ -25,6 +26,8 @@ import com.bumptech.glide.load.model.ModelLoaderFactory
 import com.bumptech.glide.load.model.MultiModelLoaderFactory
 import com.bumptech.glide.signature.ObjectKey
 import im.vector.app.core.di.ActiveSessionHolder
+import im.vector.app.core.utils.isLocalFile
+import im.vector.app.core.utils.openInputStream
 import im.vector.app.features.media.ImageContentRenderer
 import okhttp3.OkHttpClient
 import org.matrix.android.sdk.api.MatrixCallback
@@ -33,11 +36,12 @@ import java.io.File
 import java.io.IOException
 import java.io.InputStream
 
-class VectorGlideModelLoaderFactory(private val activeSessionHolder: ActiveSessionHolder)
-    : ModelLoaderFactory<ImageContentRenderer.Data, InputStream> {
+class VectorGlideModelLoaderFactory(private val context: Context,
+                                    private val activeSessionHolder: ActiveSessionHolder
+) : ModelLoaderFactory<ImageContentRenderer.Data, InputStream> {
 
     override fun build(multiFactory: MultiModelLoaderFactory): ModelLoader<ImageContentRenderer.Data, InputStream> {
-        return VectorGlideModelLoader(activeSessionHolder)
+        return VectorGlideModelLoader(context, activeSessionHolder)
     }
 
     override fun teardown() {
@@ -45,7 +49,7 @@ class VectorGlideModelLoaderFactory(private val activeSessionHolder: ActiveSessi
     }
 }
 
-class VectorGlideModelLoader(private val activeSessionHolder: ActiveSessionHolder)
+class VectorGlideModelLoader(private val context: Context, private val activeSessionHolder: ActiveSessionHolder)
     : ModelLoader<ImageContentRenderer.Data, InputStream> {
     override fun handles(model: ImageContentRenderer.Data): Boolean {
         // Always handle
@@ -53,11 +57,12 @@ class VectorGlideModelLoader(private val activeSessionHolder: ActiveSessionHolde
     }
 
     override fun buildLoadData(model: ImageContentRenderer.Data, width: Int, height: Int, options: Options): ModelLoader.LoadData<InputStream>? {
-        return ModelLoader.LoadData(ObjectKey(model), VectorGlideDataFetcher(activeSessionHolder, model, width, height))
+        return ModelLoader.LoadData(ObjectKey(model), VectorGlideDataFetcher(context, activeSessionHolder, model, width, height))
     }
 }
 
-class VectorGlideDataFetcher(private val activeSessionHolder: ActiveSessionHolder,
+class VectorGlideDataFetcher(private val context: Context,
+                             private val activeSessionHolder: ActiveSessionHolder,
                              private val data: ImageContentRenderer.Data,
                              private val width: Int,
                              private val height: Int)
@@ -97,9 +102,10 @@ class VectorGlideDataFetcher(private val activeSessionHolder: ActiveSessionHolde
 
     override fun loadData(priority: Priority, callback: DataFetcher.DataCallback<in InputStream>) {
         Timber.v("Load data: $data")
-        if (data.isLocalFile && data.url != null) {
-            val initialFile = File(data.url)
-            callback.onDataReady(initialFile.inputStream())
+        if (data.url.isLocalFile(context)) {
+            data.url.openInputStream(context)?.use {
+                callback.onDataReady(it)
+            }
             return
         }
 //        val contentUrlResolver = activeSessionHolder.getActiveSession().contentUrlResolver()